% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarometricFormula.R
\name{BarometricFormula}
\alias{BarometricFormula}
\title{Barometric height formula}
\usage{
BarometricFormula(T, h, p0 = 1013)
}
\arguments{
\item{T}{surface temperature in degrees Celsius}

\item{h}{surface elevation in [m]}

\item{p0}{reference surface pressure; defaults to 1013 mbar}
}
\value{
surface pressure in [mbar] at elevation \code{h} and temperature
\code{T}
}
\description{
Calculate atmospheric surface pressure from elevation and temperature using
the barometric height formula for an isothermal atmosphere.
}
\examples{
# Surface pressure at Kohnen Station:
BarometricFormula(-44.5, 2892)
}
\author{
Thomas Münch
}
