% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircleCoordinates.R
\name{CircleCoordinates}
\alias{CircleCoordinates}
\title{Circle coordinates on a sphere}
\usage{
CircleCoordinates(
  lat0,
  lon0,
  radius.circle,
  radius.sphere = 6357,
  n = 100,
  return.pi.interval = FALSE
)
}
\arguments{
\item{lat0}{geographical latitude [degree] of circle centre.}

\item{lon0}{geographical longitude [degree] of circle centre.}

\item{radius.circle}{radius of the circle; in the same units as
\code{radius.sphere}.}

\item{radius.sphere}{radius of the sphere on whose surface the circle is
drawn; i.e. typically Earth's radius. Default value is Earth's polar radius
which gives slightly more accurate results for drawing circles close to
either one of the poles. Adjust to equatorial radius for drawing circles in
the mid-latitudes.}

\item{n}{number of polar angle increments to draw the circle.}

\item{return.pi.interval}{if \code{TRUE}, return the circle's longitudes in
the interval [-pi, pi]; else in [0, 2*pi].}
}
\value{
data frame of circle latitude (\code{lat}) and longitudes
(\code{lon}) in geographical coordinates in degree for the number of
incremental steps specified by \code{n} (given as the polar coordinate
\code{phi}).
}
\description{
Get the latitude and longitude coordinates of a circle with a given radius
drawn onto the surface of a sphere.
}
\author{
Thomas Münch
}
